reorganize board files
