+++ include/boards/avr/emoro_variants.h
