+++ include/boards/avr/ethernet.h
