+++ include/boards/samd/adi.h
